requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.4
tzdata==2025.2

# Database
psycopg2-binary==2.9.10

# Environment variables
python-dotenv==1.0.0

# Optional but recommended
django-cors-headers==4.6.0
Pillow==11.0.0

